Remove GAMMA (proprietary) script and hack
mk_geo_radcal is a pearl script provided by GAMMA. The version in this
repository is a hacked version of circa 2015 GAMMA and (per Tom) the
hacks have likely been depreciated.

Kirk indicated that these were likely done at GAMMA's suggestion, and
including this was done at GAMMA's direction, but, it's better to not
include any proprietary code in this repo, so removal is well needed.

If it turns out the hack is still needed, some other method should be
used to provide the hack without including proprietary code, OR
documentation should provided that this is okay.