Move execute (subprocess) commands inside execute statement
The useage of execute, which subprocesses commands out, has typically
been:
   cmd = "SOME COMMAND"
   execute(cmd)
However, this usage pattern makes it hard to determine the underlying
system dependencies needed. Moving cmd inside execute, and ensuring the
command executed in the shell is on the same line, makes it possible to
grep for external dependencies.