Setup a base Jupyter prototyping environment
This transforms our docker to a multistage build, where our production
containers are build by targeting "production" `--target production`,
and if no target is (or the prototype target is) specified, a
prototyping environment is built into the container.

Docker run command should look like:
    docker run docker run --rm \
        -v ${PWD}:/home/conda/hyp3-gamma -p 8888:8888 -it ...

assuming you're in the repository root.

This prototyping environment is based on:
https://github.com/jupyter/docker-stacks/tree/master/base-notebook